Fixes startup smsg poll when DB is empty. When starting up, the core rpc command 'smsgzmqpush' is used to poll the DB to send any missed smsg messages. However, such a poll is based on the last date that an smsg message occurred, but if the DB has no previous smsg messages then such a request is never made.
This change fixes this to then instead use the maximum duration possible instead to determine from when to request messages from core: a paid message duration.
Note that in the case of duplicate messages, the DB already has a unique constraint on the msgid and direction of the message, which correctly fails for duplicate messages found, and thus the case for potential duplicated messages requires no additional handling with this change.